Enforce site scope in Manage Users / Manage Groups administration flows

+++ src/main/java/org/jahia/modules/sitesettings/users/UsersFlowHandler.java
@@ -66,6 +66,19 @@ public class UsersFlowHandler implements Serializable {
         }
     }
 
+    /**
+     * When the management screen is bound to a specific site (siteKey set from a jnt:virtualsite realm),
+     * a target principal is only in scope if it is stored under that site's principal tree
+     * (/sites/&lt;siteKey&gt;/...). In the server-wide realm (siteKey null) there is no site restriction.
+     * This mirrors the store layout used by JahiaUserManagerService ("/users/" vs "/sites/&lt;siteKey&gt;/users/").
+     */
+    private boolean isInAdministeredScope(String principalPath) {
+        if (siteKey == null) {
+            return true;
+        }
+        return principalPath != null && principalPath.startsWith("/sites/" + siteKey + "/");
+    }
+
     public boolean addUser(final UserProperties userProperties, final MessageContext context) throws RepositoryException {
         logger.info("Adding user");
         return JCRTemplate.getInstance().doExecuteWithSystemSession(new JCRCallback<Boolean>() {
@@ -216,6 +229,11 @@ public class UsersFlowHandler implements Serializable {
             @Override
             public Boolean doInJCR(JCRSessionWrapper session) throws RepositoryException {
                 JCRUserNode jahiaUser = userManagerService.lookupUserByPath(userKey);
+                if (jahiaUser == null || !isInAdministeredScope(jahiaUser.getPath())) {
+                    context.addMessage(new MessageBuilder().error().code(
+                            "siteSettings.user.remove.unsuccessful").arg(userKey).build());
+                    return false;
+                }
                 String displayName = PrincipalViewHelper.getDisplayName(jahiaUser);
                 if (userManagerService.deleteUser(jahiaUser.getPath(), session)) {
                     context.addMessage(new MessageBuilder().info().code(
@@ -289,6 +307,11 @@ public class UsersFlowHandler implements Serializable {
             @Override
             public Boolean doInJCR(JCRSessionWrapper session) throws RepositoryException {
                 JCRUserNode jahiaUser = userManagerService.lookupUserByPath(userProperties.getUserKey(), session);
+                if (jahiaUser == null || !isInAdministeredScope(jahiaUser.getPath())) {
+                    context.addMessage(new MessageBuilder().error().source("userName").code(
+                            "siteSettings.user.edit.errors.property").arg("userName").build());
+                    return false;
+                }
                 boolean hasErrors = false;
                 Set<String> readOnlyProps = userProperties.getReadOnlyProperties();
                 if (jahiaUser != null) {
